requirements.txt
# Core ML Libraries
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP
nltk>=3.8.1

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging and Experiment Tracking
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Data Download
gdown>=4.7.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.8.0
dvc>=3.0.0
dvc-s3>=3.0.0  # For S3 remote (optional)
dvc-gs>=3.0.0  # For GCS remote (optional)
dvc-azure>=3.0.0  # For Azure remote (optional)

# Hyperparameter Tuning
optuna>=3.3.0
optuna-integration>=3.0.0

# API Development
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0  # For testing

# Monitoring
scipy>=1.11.0  # For statistical tests in drift detection

